The frozen job handler has been changed again.
I have found that trying to be to clever with the frozen scf part of the
check method has leads to too many headaches. While it was a nice idea,
it doesn't seem to work. The last idea was to count scf loops to see if
the calc was in the middle of scf, but it failed when the overall scf, but
the time in between the scf and finishing/proceeding to next scf was longer
than scf timeout.