NonConvergingErrorHandler for meta-GGAs/hybrids
Don't do VeryFast --> Fast --> Normal --> All for meta-GGAs hybrids. For meta-GGAs, use Algo = All if it wasn't set; otherwise, that's it. For hybrid GGAs, set Algo = All if it wasn't set; otherwise, one could consider Damped with TIME = 0.5 per the VASP manual, but hybrids are so expensive this may not be worthwhile (commented out for now).