More appropriate grad_not_orth fix
grad_not_orth is primarily an issue related to how VASP is compiled. Instead of changing ISMEAR, Custodian now checks to see if Algo = All or Algo = Damped is set, in which case Algo = Normal is the new algorithm. Algo = All is *not* modified if METAGGA = True or LHFCALC = True.