More robust zbrent fix
zbrent often occurs when you are near the PES minimum and the numerical noise in the forces is comparatively large. This commit more robustly addresses this by tightening EDIFF and increasing NELMIN if CONTCAR --> POSCAR on its own doesn't help. If VTST fixes are supported, FIRE is almost a guarantee success and is applied instead of IBRION = 1.